+++ fedoo/weakform/stress_equilibrium_mixed.py
@@ -31,6 +31,9 @@ class StressEquilibriumMixed(StressEquilibrium):
         Bulk modulus used for scaling the pressure equation. If None, it
         is estimated from the tangent matrix trace.
         Note: This is used for scaling the pressure constraint equation.
+    convert_tangent: bool, default=True
+        Convert the constitutive corotational Kirchhoff tangent to the
+        formulation tangent.
     space: ModelingSpace, optional
         Modeling space.
 
@@ -47,12 +50,14 @@ class StressEquilibriumMixed(StressEquilibrium):
         self,
         constitutivelaw,
         bulk_modulus=None,
+        convert_tangent=True,
         name="",
         nlgeom=None,
         space=None,
     ):
         super().__init__(
             constitutivelaw,
+            convert_tangent=convert_tangent,
             name=name,
             nlgeom=nlgeom,
             space=space,